实现接口 /sell/buyer/product/list

+++ src/main/java/com/example/orderingsystem/repository/SellerInfoRepository.java
@@ -3,7 +3,7 @@ package com.example.orderingsystem.repository;
 import com.example.orderingsystem.entity.SellerInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SellerInfoRepository extends JpaRepository<SellerInfo, Integer> {
+public interface SellerInfoRepository extends JpaRepository<SellerInfo, Long> {
 
 
 }
