+++ src/main/java/com/example/orderingsystem/service/SellerService.java
@@ -1,11 +1,10 @@
 package com.example.orderingsystem.service;
 
 import com.example.orderingsystem.entity.SellerInfo;
-import org.springframework.stereotype.Service;
 
 public interface SellerService {
 
-	SellerInfo findOne(Integer sellerId);
+	SellerInfo findOne(Long sellerId);
 
 	SellerInfo save(SellerInfo sellerInfo);
 }
